file resturcture and added functions
